feature[Req5-5.2]:Consultar Ordenes

+++ OrdersMS.Domain/Entities/Tarifa.cs
@@ -10,7 +10,7 @@ namespace OrdersMS.Domain.Entities
     {
         public Guid Id { get; private set; }
         public string Nombre { get; private set; }
-        public decimal CostoBase { get;  private set; }
+        public double CostoBase { get;  private set; }
         public double DistanciaKm { get;  private set; }
         public double CostoPorKm { get;  private set; }
         public string Estatus { get; private  set; }
